+++ python/tvm/__init__.py
@@ -121,7 +121,7 @@ sys.excepthook = tvm_wrap_excepthook(sys.excepthook)
 # Autoload out-of-tree backends registered under the ``tvm.backends`` entry
 # point group. Runs last, after the core runtime and the tvm namespace are
 # fully initialized, so an extension can safely register into ``tvm.*`` and
-# load extra libraries. Imported lazily here to avoid any import-cycle risk.
-from ._autoload_backends import _autoload_backends
+# load extra libraries.
+from .backend._autoload_backends import _autoload_backends
 
 _autoload_backends()
